fix: further code clean and fix
