+++ crates/iota-indexer/Cargo.toml
@@ -84,6 +84,7 @@ iota-move-build.workspace = true
 iota-swarm-config.workspace = true
 iota-test-transaction-builder.workspace = true
 simulacrum.workspace = true
+simulacrum-server.workspace = true
 test-cluster.workspace = true
 
 [[bin]]
@@ -91,4 +92,4 @@ name = "iota-indexer"
 path = "src/main.rs"
 
 [package.metadata.cargo-udeps.ignore]
-development = ["simulacrum"]
+development = ["simulacrum", "simulacrum-server"]
