+++ crates/simulacrum-server/Cargo.toml
@@ -32,6 +32,7 @@ uuid.workspace = true
 iota-config.workspace = true
 iota-execution.workspace = true
 iota-faucet.workspace = true
+iota-grpc-server.workspace = true
 iota-node-storage.workspace = true
 iota-json-rpc-types.workspace = true
 iota-protocol-config.workspace = true
