+++ crates/simulacrum-server/src/grpc_server.rs
@@ -1,9 +1,46 @@
 // Copyright (c) 2026 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! gRPC streaming server for Simulacrum
+//! gRPC server utilities for Simulacrum.
+//!
+//! Provides [`start_simulacrum_grpc_server`] which wires up a complete gRPC
+//! server backed by a [`Simulacrum`] instance, including proper chain ID
+//! resolution and an optional [`TransactionExecutor`].
 
-pub use simulacrum::grpc::start_simulacrum_grpc_server;
+use std::sync::Arc;
+
+use anyhow::Result;
+use iota_grpc_server::{GrpcReader, GrpcServerHandle, start_grpc_server};
+use iota_node_storage::GrpcStateReader;
+use iota_types::transaction_executor::TransactionExecutor as TransactionExecutorTrait;
+use simulacrum::{Simulacrum, transaction_executor::TransactionExecutor};
+use tokio_util::sync::CancellationToken;
+
+/// Start a gRPC server for the given simulacrum instance.
+///
+/// This creates the full server stack with:
+/// - The [`Simulacrum`] as the state reader (implements [`GrpcStateReader`])
+/// - A [`TransactionExecutor`] for transaction execution/simulation
+/// - The real chain identifier from the simulacrum genesis
+pub async fn start_simulacrum_grpc_server(
+    simulacrum: Arc<Simulacrum>,
+    config: iota_config::node::GrpcApiConfig,
+    shutdown_token: CancellationToken,
+) -> Result<GrpcServerHandle> {
+    let chain_id = simulacrum
+        .get_chain_identifier()
+        .expect("chain identifier should be set");
+
+    // Create a transaction executor for simulacrum to enable transaction execution
+    // and simulation via gRPC
+    let executor =
+        Some(Arc::new(TransactionExecutor::new(simulacrum.clone()))
+            as Arc<dyn TransactionExecutorTrait>);
+
+    let grpc_reader = Arc::new(GrpcReader::new(simulacrum, None));
+
+    start_grpc_server(grpc_reader, executor, config, shutdown_token, chain_id).await
+}
 
 #[cfg(test)]
 mod tests {
