+++ crates/simulacrum/Cargo.toml
@@ -13,13 +13,11 @@ async-trait.workspace = true
 fastcrypto.workspace = true
 prometheus.workspace = true
 rand.workspace = true
-tokio-util.workspace = true
 tracing.workspace = true
 
 # internal dependencies
 iota-config.workspace = true
 iota-execution.workspace = true
-iota-grpc-server.workspace = true
 iota-node-storage.workspace = true
 iota-protocol-config.workspace = true
 iota-storage.workspace = true
